# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
        "/api/habitaciones": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener lista de habitaciones paginada con filtros dinámicos",
                "operationId": "be9f776a8011ae0db75dbc6963e4281a",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "query",
                        "description": "Número de habitación para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Tipo de habitación para filtrar (ej: simple, doble)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "precioNoche",
                        "in": "query",
                        "description": "Precio por noche para filtrar",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "ID del hotel asociado para filtrar",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de habitaciones filtrada y paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Crear una nueva habitación",
                "operationId": "756c9b19b774c3a42b0dadefd5aee62e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHabitacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Habitación creada correctamente"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            }
        },
...
        "/api/habitaciones/{habitacion}": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener detalles de una habitación",
                "operationId": "907b4b2bdef244c31e212d32e44c7edd",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la habitación"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Actualizar una habitación",
                "operationId": "1eae9de9b0cb5515936b87a34cde18de",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHabitacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Habitación actualizada correctamente"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Eliminar una habitación",
                "operationId": "d19fc67d831c97d432489dbf0de10de2",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación eliminada correctamente"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Actualizar parcial de una habitación",
                "operationId": "6fa5f5c872e25009721ccc1de45e6bb2",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHabitacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Habitación actualizada correctamente"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            }
        },
        "/api/habitaciones/{habitacion}/hotel": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener el hotel asociado a una habitación",
                "operationId": "95a895a8f86f313d1e4019ed221a8da9",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel asociado"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            }
        },
        "/api/habitaciones/{habitacion}/huespedes": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener los huéspedes de una habitación",
                "operationId": "30a21f1063309b641041756d6b8bc76f",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de huéspedes"
                    },
                    "404": {
                        "description": "Habitación no encontrada o sin huéspedes"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreHabitacionRequest": {
                "title": "Store Habitación Request",
                "description": "Datos requeridos para crear una nueva habitación",
                "required": [
                    "numero",
                    "tipo",
                    "precioNoche",
                    "hotel_id"
                ],
                "properties": {
                    "numero": {
                        "description": "Número de la habitación",
                        "type": "string",
                        "maxLength": 10
                    },
                    "tipo": {
                        "description": "Tipo de habitación",
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 5
                    },
                    "precioNoche": {
                        "description": "Precio por noche",
                        "type": "number",
                        "format": "float"
                    },
                    "hotel_id": {
                        "description": "ID del hotel al que pertenece",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PutHabitacionRequest": {
                "title": "Put Habitación Request",
                "description": "Datos requeridos para actualizar una habitación",
                "required": [
                    "numero",
                    "tipo",
                    "precioNoche",
                    "hotel_id"
                ],
                "properties": {
                    "numero": {
                        "description": "Número de la habitación",
                        "type": "string",
                        "maxLength": 10
                    },
                    "tipo": {
                        "description": "Tipo de habitación",
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 5
                    },
                    "precioNoche": {
                        "description": "Precio por noche",
                        "type": "number",
                        "format": "float"
                    },
                    "hotel_id": {
                        "description": "ID del hotel al que pertenece",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}